+++ src/main/java/io/kloudfile/telegram/bot/bots/Bot.java
@@ -1,6 +1,6 @@
 package io.kloudfile.telegram.bot.bots;
 
-import io.kloudfile.telegram.bot.dto.callbackDTO.ResponseDTO;
+import io.kloudfile.telegram.bot.dto.infosys.callbackDTO.ResponseDTO;
 
 import java.util.List;
