+++ src/main/java/io/kloudfile/telegram/bot/dto/infosys/Chat.java
@@ -1,31 +1,26 @@
 
-package io.kloudfile.telegram.bot.dto.callbackDTO;
+package io.kloudfile.telegram.bot.dto.infosys;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-public class From {
+public class Chat {
 
-    @SerializedName("last_name")
-    @Expose
-    private String lastName;
     @SerializedName("id")
     @Expose
     private Integer id;
     @SerializedName("first_name")
     @Expose
     private String firstName;
+    @SerializedName("last_name")
+    @Expose
+    private String lastName;
     @SerializedName("username")
     @Expose
     private String username;
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
+    @SerializedName("type")
+    @Expose
+    private String type;
 
     public Integer getId() {
         return id;
@@ -43,12 +38,28 @@ public class From {
         this.firstName = firstName;
     }
 
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
     public String getUsername() {
         return username;
     }
 
     public void setUsername(String username) {
         this.username = username;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
     }
 
 }
