+++ src/main/java/io/kloudfile/telegram/bot/dto/infosys/callbackDTO/ResponseDTO.java
@@ -1,5 +1,5 @@
 
-package io.kloudfile.telegram.bot.dto.callbackDTO;
+package io.kloudfile.telegram.bot.dto.infosys.callbackDTO;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
